updating repo classes

+++ ProjectCinderella.Data/Components/CrudComponents/DeleteEntityComponent.cs
@@ -4,9 +4,6 @@ namespace ProjectCinderella.Data.Components.CrudComponents
 {
     public class DeleteEntityComponent
     {
-        public void Execute<T>(IRepository<T> repo, int id, string userID) where T : class
-        {
-            repo.Delete(id, userID);
-        }
+        public void Execute<T>(IRepository<T> repo, int id, string userID) where T : class => repo.Delete(id, userID);
     }
 }
